requirements.txt
fastapi
uvicorn
praw
python-dotenv
requests
jinja2
aiofiles